feat: added 'add new category'
