+++ frontend/index.html
@@ -36,8 +36,6 @@
         <select id="to"></select>
         Category:
         <select id="category"></select>
-        <input type="text" />
-        <button type="button">Add new category</button>
 
         Description
         <input type="text" />
@@ -47,7 +45,10 @@
 
         <button>Add transaction</button>
       </form>
-
+      <form id="create-new-category">
+        <input type="text" />
+        <button type="submit">Add new category</button>
+      </form>
       Account summary
       <ul></ul>
       Filter by account:
